+++ src/PhinxAbstractSeed.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FiiSoft\Phinx;
+
+use Phinx\Seed\AbstractSeed;
+
+abstract class PhinxAbstractSeed extends AbstractSeed
+{
+    use PhinxCommonTrait;
+}