+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All important changes to `fiisoft-phinx` will be documented in this file
 
+## 1.3.0
+
+* added new abstract class PhinxAbstractSeed
+* modified command PhinxSeedCreateCmd
+* added trait PhinxCommonTrait
+* changed class PhinxConfig
+* readme updated with example of the simplest configuration 
+* some other fixes and improvements
+
 ## 1.2.0
 
 Fixed bug with dropView and added new method hasView() to PhinxAbstractMigration.